move test things to conftest.py
